flac: Drop "--disable-asm-optimization"
It's actually a typo for "--disable-asm-optimizations" (note the
trailing "s").  It was ineffective in previous versions, indicating it's
not really needed.  Fedora and LFS do not have this as well.  And now
in 1.5.0 the typo causes a hard error.

I've also read the code, this switch only has an effect for x86 and
aarch64.  On x86 it uses CPUID at runtime to probe the hardware feature
thus the asm optimization shouldn't cause any trouble.  On aarch64 it
only uses neon, that aarch64 mandates.