Revert "loonggl: Apply a mystical binary patch to fix error handling"
This reverts commit 3136766e85b0ac14711ef85221eb558400678516.

The error handling added by the mystical patch just prevents the crash,
but it still fails to allow libglvnd to fall back to Mesa.  When Xorg is
started the EGL-based glamor acceleration will actually fail to
initialize, and in a X session running any EGL program will also fail.

A different workaround will be attempted in the following commit(s).